requirements.txt
# Core Framework
streamlit==1.29.0

# LLM and AI
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1
huggingface-hub==0.20.0

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Vector Store
chromadb==0.4.22

# Embeddings
sentence-transformers==2.2.2

# Environment Variables
python-dotenv==1.0.0

# API Framework (optional - for future REST API)
fastapi==0.108.0
uvicorn==0.25.0

# Utilities
tiktoken==0.5.2
